pyproject.toml
[tool.poetry]
name = "renterx"
version = "0.1.0"
description = "RenterX Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
factory-boy = "^3.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
ruff = "^0.1.14"
pytest = "^8.0.0"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"